fix(bench): preload parked/cold drives before reading UFFS record counts
Parked drives appear in 'uffs daemon status' without a record count:
  [Parked] G: — bloom + trie kept resident; body released

warm_parked_drives() fires 'uffs daemon preload <DRIVE>' for each
candidate drive absent from the initial warm-map, then re-reads
'uffs daemon status' once so the drive appears as [Warm] with a live
record count.  Preload failure is non-fatal (count stays 0).

- Add warm_parked_drives() helper in preflight/mod.rs
- capture() calls it between first status read and probe_drive loop;
  second status read only runs when at least one drive was absent
- Add DAEMON_READY_STATUS C Warm line in run/tests.rs so the C test
  drive is already warm (no preload injected there)
- Add parked_drive_is_preloaded_and_count_populated test
- Update unconfigured_drive_probed_once_without_sleep: mock preload +
  second status; assert 5 Run calls instead of 3
- Extract preflight test module to preflight/tests.rs (mod.rs 559 LOC,
  tests.rs 295 LOC — both under 800 LOC policy limit)
